+++ src/electron-start.js
@@ -9,6 +9,7 @@ const windowStateKeeper = require('electron-window-state')
 // be closed automatically when the JavaScript object is garbage collected.
 let win, winTrial, winSerial
 let changed = false
+let colNumber = 1
 const hiddenStore = new Store({name: "kernel", encryptionKey: "79df64f73eab9bc0d7b448d2008d876e"})
 const userStore = new Store({name: "config"})
 
@@ -91,6 +92,37 @@ function validSerial(email, storedSerial) {
 }
 
 
+function exportMenu(cols) {
+  var expMenu =
+    [ { label : 'Entire Document...'
+      , click (item, focusedWindow) {
+          focusedWindow.webContents.send('menu-export-txt')
+        }
+      }
+    , { label : 'Current Card and Children...'
+      , click (item, focusedWindow) {
+          focusedWindow.webContents.send('menu-export-txt-current')
+        }
+      }
+    , { type: 'separator' }
+    ]
+
+  var expMenuItem = function (num) {
+    return  { label : `Column ${num}...`
+            , click (item, focusedWindow) {
+                focusedWindow.webContents.send('menu-export-txt-column', num)
+              }
+            }
+  }
+
+  for (var i = 1; i <= cols;i++) {
+    expMenu.push(expMenuItem(i))
+  }
+
+  return expMenu
+}
+
+
 
 /* ==== App Events ==== */
 
@@ -138,6 +170,13 @@ app.on('activate', () => {
 
 ipcMain.on('changed', (event, msg) => {
   changed = msg;
+})
+
+
+ipcMain.on('column-number-change', (event, msg) => {
+  menuTemplate = menuFunction(msg)
+  menu = Menu.buildFromTemplate(menuTemplate)
+  Menu.setApplicationMenu(menu)
 })
 
 
@@ -219,8 +258,8 @@ function createSerialWindow(shouldBlock) {
 
 /* ==== Menu ==== */
 
-const menuTemplate =
-  [ { label: 'File'
+function menuFunction(cols) {
+  return [ { label: 'File'
     , submenu:
         [ { label: 'New'
           , accelerator: 'CmdOrCtrl+N'
@@ -259,10 +298,8 @@ const menuTemplate =
               focusedWindow.webContents.send('menu-export-json')
             }
           }
-        , { label: 'Export Text File...'
-          , click (item, focusedWindow) {
-              focusedWindow.webContents.send('menu-export-txt')
-            }
+        , { label: 'Export Text'
+          , submenu : exportMenu(cols)
           }
         , { type: 'separator' }
         , { label: 'Exit...'
@@ -336,6 +373,9 @@ const menuTemplate =
         ]
     }
   ]
+}
+
+var menuTemplate = menuFunction(colNumber)
 
 if(true) {//process.defaultApp || /[\\/]electron-prebuilt[\\/]/.test(process.execPath) || /[\\/]electron[\\/]/.test(process.execPath)) {
   menuTemplate.push(
@@ -352,4 +392,4 @@ if(true) {//process.defaultApp || /[\\/]electron-prebuilt[\\/]/.test(process.exe
  )
 }
 
-const menu = Menu.buildFromTemplate(menuTemplate)
+var menu = Menu.buildFromTemplate(menuTemplate)